+++ src/assets/svg/IcSignupOwner.tsx
@@ -1,0 +1,11 @@
+import * as React from "react";
+import type { SVGProps } from "react";
+const SvgIcSignupOwner = (props: SVGProps<SVGSVGElement>) => (
+  <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 80 80" {...props}>
+    <path
+      fill="#3BA8FF"
+      d="M12.566 33.625h54.868c1.967 0 3.566-1.644 3.566-3.666 0-.727-.213-1.434-.6-2.032l-7.316-11.295c-.862-1.335-2.306-2.132-3.866-2.132H20.791a4.61 4.61 0 0 0-3.865 2.132L9.6 27.917A3.8 3.8 0 0 0 9 29.95c0 2.032 1.599 3.676 3.566 3.676m2.635 3.188v23.906c0 2.64 2.083 4.781 4.65 4.781h21.704c2.568 0 4.65-2.142 4.65-4.781V36.813h-6.2V52.75H21.402V36.813zm43.407 0v25.5c0 1.763 1.385 3.187 3.1 3.187s3.1-1.424 3.1-3.187v-25.5z"
+    />
+  </svg>
+);
+export default SvgIcSignupOwner;
