+++ src/assets/svg/IcSignupUser.tsx
@@ -1,0 +1,11 @@
+import * as React from "react";
+import type { SVGProps } from "react";
+const SvgIcSignupUser = (props: SVGProps<SVGSVGElement>) => (
+  <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 80 80" {...props}>
+    <path
+      fill="#3BA8FF"
+      d="M21.3 13c2.254 0 4.416.889 6.01 2.471a8.4 8.4 0 0 1 2.49 5.967 8.4 8.4 0 0 1-2.49 5.966 8.53 8.53 0 0 1-6.01 2.471 8.53 8.53 0 0 1-6.01-2.471 8.4 8.4 0 0 1-2.49-5.966c0-2.238.895-4.384 2.49-5.967A8.53 8.53 0 0 1 21.3 13m39.1 0c2.254 0 4.416.889 6.01 2.471a8.4 8.4 0 0 1 2.49 5.967 8.4 8.4 0 0 1-2.49 5.966 8.53 8.53 0 0 1-6.01 2.471 8.53 8.53 0 0 1-6.01-2.471 8.4 8.4 0 0 1-2.49-5.966c0-2.238.895-4.384 2.49-5.967A8.53 8.53 0 0 1 60.4 13M6 44.504C6 38.29 11.079 33.25 17.337 33.25h4.537c1.69 0 3.294.37 4.739 1.023a13 13 0 0 0-.202 2.352c0 4.029 1.785 7.647 4.6 10.125H8.263C7.02 46.75 6 45.737 6 44.504m43.063 2.246h-.074c2.826-2.478 4.6-6.096 4.6-10.125 0-.802-.074-1.582-.201-2.352a11.3 11.3 0 0 1 4.738-1.023h4.537C68.921 33.25 74 38.291 74 44.504a2.257 2.257 0 0 1-2.263 2.246H49.064M29.8 36.625c0-2.685 1.075-5.26 2.988-7.16A10.24 10.24 0 0 1 40 26.5c2.705 0 5.3 1.067 7.212 2.965a10.09 10.09 0 0 1 2.988 7.16c0 2.685-1.075 5.26-2.988 7.16A10.24 10.24 0 0 1 40 46.75c-2.705 0-5.3-1.067-7.212-2.965a10.09 10.09 0 0 1-2.988-7.16M19.6 64.184c0-7.762 6.343-14.059 14.163-14.059h12.463c7.83 0 14.174 6.297 14.174 14.059A2.826 2.826 0 0 1 57.563 67H22.426c-1.562 0-2.837-1.255-2.837-2.816z"
+    />
+  </svg>
+);
+export default SvgIcSignupUser;
