+++ src/routes/Router.tsx
@@ -18,6 +18,7 @@ import {
   UsePass,
   Signin,
   Signup,
+  SignupStep1,
   SignupStep2,
   ChangePhone,
   Payment,
@@ -57,7 +58,7 @@ const Router = createBrowserRouter([
     element: (
       <>
         <ScrollToTop />
-        <AuthHandler /> {/* ✅ 소셜 로그인 후 자동 실행 */}
+        <AuthHandler /> {/* 소셜 로그인 후 자동 실행 */}
         <RootLayout />
       </>
     ),
@@ -72,10 +73,14 @@ const Router = createBrowserRouter([
         element: <Signup />,
       },
       {
+        path: "/signup/step1",
+        element: <SignupStep1 />,
+      },
+      {
         path: "/signup/step2",
         element: (
           <>
-            <AuthHandler /> {/* ✅ 소셜 로그인 후 자동 실행 */}
+            <AuthHandler /> {/* 소셜 로그인 후 자동 실행 */}
             <SignupStep2 />
           </>
         ),
